+++ src/modules/usuario/usuario.controller.ts
@@ -35,8 +35,17 @@ export async function getById(req: Request, res: Response) {
 export async function update(req: Request, res: Response) {
   try {
     const id = Number(req.params.id);
+
+
+    const usuario = await usuarioService.getUsuarioById(id);
+    if (!usuario) return res.status(404).json({ mensagem: 'Usuário não encontrado' });
+
+    const usuarioAtualizado = await usuarioService.updateUsuario(id, req.body);
+    return res.status(200).json({  mensagem: "Usuário atualizado com sucesso", data: usuarioAtualizado });
+
     const usuario = await usuarioService.updateUsuario(id, req.body);
     return res.status(200).json({  sucesso: true, mensagem: "Usuário atualizado com sucesso", data: usuario });
+
   } catch (error) {
     return res.status(400).json({ sucesso: false, mensagem: "Erro ao atualizar usuário", erro: (error as Error).message });
   }
@@ -45,6 +54,10 @@ export async function update(req: Request, res: Response) {
 export async function remove(req: Request, res: Response) {
   try {
     const id = Number(req.params.id);
+
+    const usuario = await usuarioService.getUsuarioById(id);
+    if (!usuario) return res.status(404).json({ mensagem: 'Usuário não encontrado' });
+
     await usuarioService.deleteUsuario(id);
     return res.status(204).json({ sucesso: true, mensagem: "Usuário deletado com sucesso" });
   } catch (error) {
